YaTool : tests : python2 -> python3, strin -> bytes and some fixes
